+++ samples/collision/world.cpp
@@ -137,8 +137,8 @@ void World :: OnCollision( SunLight :: Collision :: Collider *pFirst, SunLight :
 
 bool World :: LoadSprites( void ) {
 
-    if( m_pCanvasSunny -> Load( m_strBasePath + __SUNNY_SPRITE_IDLE ) &&
-        m_pCanvasObstacle -> Load( m_strBasePath + __OBSTACLE_IMAGE ) ) {
+    if( m_pCanvasSunny -> Load( __SUNNY_SPRITE_IDLE ) &&
+        m_pCanvasObstacle -> Load( __OBSTACLE_IMAGE ) ) {
         SunLight :: TileMap :: stDimension2D    dimSunny;
         SunLight :: TileMap :: stDimension2D    dimObstacle;
 
@@ -211,7 +211,6 @@ bool World :: LoadSprites( void ) {
  */
 World :: World( std :: string strBasePath )  {
 
-    m_strBasePath = strBasePath;
     m_LastStep     = { 0, 0 };
     m_pRenderer = std :: make_unique<SunLight :: Renderer :: TileMapRenderer>( __DISPLAY_W,
                                                                                __DISPLAY_H,
@@ -260,7 +259,7 @@ bool World :: Run( void )  {
     m_pRenderer -> SetDrawFPS( __ENABLE_FPS_SHOW_LABEL );
     m_pRenderer -> Start();
 
-    strMapFile = m_strBasePath + __TMX_MAP_FILE;
+    strMapFile = __TMX_MAP_FILE;
 
     if( !m_pRenderer -> LoadMap( strMapFile.c_str(), __DEFAULT_MAP_ALIGNMENT ) )  {
         fprintf( stderr, "Error loading map\n" );